[Security Solution] Enable actions in document details preview footer (#203691)
## Summary

Updated document details preview footer to also show actions
### Before
<img width="1481" alt="image"
src="https://github.com/user-attachments/assets/aace12aa-f3d1-4fe2-a3b7-392878207f39"
/>

### After
- Users can perform alert/event actions in a preview

![image](https://github.com/user-attachments/assets/d42be26d-9d4a-4701-bc88-92549ebfb65c)

- In analyzer, when examining an event, event actions are also available

![image](https://github.com/user-attachments/assets/d30515d9-b428-4112-86f8-9bb872eaf921)

- No change to flyout in rule creation workflow, action is not available
in preview nor non-preview

### Checklist

- [x] [Unit or functional
tests](https://www.elastic.co/guide/en/kibana/master/development-tests.html)
were updated or added to match the most common scenarios

(cherry picked from commit 7aac71c2aa5eb78f57bfe7bbb047104766b55a30)
