+++ x-pack/plugins/security_solution/public/flyout/document_details/preview/footer.test.tsx
@@ -14,24 +14,58 @@ import { mockContextValue } from '../shared/mocks/mock_context';
 import { DocumentDetailsContext } from '../shared/context';
 import { PreviewPanelFooter } from './footer';
 import { PREVIEW_FOOTER_TEST_ID, PREVIEW_FOOTER_LINK_TEST_ID } from './test_ids';
+import { FLYOUT_FOOTER_DROPDOWN_BUTTON_TEST_ID } from '../shared/components/test_ids';
 import { createTelemetryServiceMock } from '../../../common/lib/telemetry/telemetry_service.mock';
+import { useKibana } from '../../../common/lib/kibana';
+import { useAlertExceptionActions } from '../../../detections/components/alerts_table/timeline_actions/use_add_exception_actions';
+import { useInvestigateInTimeline } from '../../../detections/components/alerts_table/timeline_actions/use_investigate_in_timeline';
+import { useAddToCaseActions } from '../../../detections/components/alerts_table/timeline_actions/use_add_to_case_actions';
 
 jest.mock('@kbn/expandable-flyout');
-
-const mockedTelemetry = createTelemetryServiceMock();
-jest.mock('../../../common/lib/kibana', () => {
+jest.mock('react-router-dom', () => {
+  const original = jest.requireActual('react-router-dom');
   return {
-    useKibana: () => ({
-      services: {
-        telemetry: mockedTelemetry,
-      },
-    }),
+    ...original,
+    useLocation: jest.fn().mockReturnValue({ search: '' }),
   };
 });
 
+jest.mock('../../../common/lib/kibana');
+jest.mock('../../../detections/components/alerts_table/timeline_actions/use_add_exception_actions');
+jest.mock(
+  '../../../detections/components/alerts_table/timeline_actions/use_investigate_in_timeline'
+);
+jest.mock('../../../detections/components/alerts_table/timeline_actions/use_add_to_case_actions');
+
+const mockedTelemetry = createTelemetryServiceMock();
+
 describe('<PreviewPanelFooter />', () => {
-  beforeAll(() => {
+  beforeEach(() => {
     jest.mocked(useExpandableFlyoutApi).mockReturnValue(mockFlyoutApi);
+    (useKibana as jest.Mock).mockReturnValue({
+      services: {
+        osquery: { isOsqueryAvailable: jest.fn() },
+        telemetry: mockedTelemetry,
+        cases: { hooks: { useIsAddToCaseOpen: jest.fn().mockReturnValue(false) } },
+      },
+    });
+    (useAlertExceptionActions as jest.Mock).mockReturnValue({ exceptionActionItems: [] });
+    (useInvestigateInTimeline as jest.Mock).mockReturnValue({
+      investigateInTimelineActionItems: [],
+    });
+    (useAddToCaseActions as jest.Mock).mockReturnValue({ addToCaseActionItems: [] });
+  });
+
+  it('should not render the take action dropdown if preview mode', () => {
+    const { queryByTestId } = render(
+      <TestProviders>
+        <DocumentDetailsContext.Provider value={{ ...mockContextValue, isPreview: true }}>
+          <PreviewPanelFooter />
+        </DocumentDetailsContext.Provider>
+      </TestProviders>
+    );
+
+    expect(queryByTestId(PREVIEW_FOOTER_TEST_ID)).not.toBeInTheDocument();
   });
 
   it('should render footer for alert', () => {
@@ -43,7 +77,7 @@ describe('<PreviewPanelFooter />', () => {
       </TestProviders>
     );
     expect(getByTestId(PREVIEW_FOOTER_TEST_ID)).toBeInTheDocument();
-    expect(getByTestId(PREVIEW_FOOTER_TEST_ID)).toHaveTextContent('Show full alert details');
+    expect(getByTestId(PREVIEW_FOOTER_LINK_TEST_ID)).toHaveTextContent('Show full alert details');
   });
 
   it('should render footer for event', () => {
@@ -56,7 +90,21 @@ describe('<PreviewPanelFooter />', () => {
         </DocumentDetailsContext.Provider>
       </TestProviders>
     );
-    expect(getByTestId(PREVIEW_FOOTER_TEST_ID)).toHaveTextContent('Show full event details');
+    expect(getByTestId(PREVIEW_FOOTER_LINK_TEST_ID)).toHaveTextContent('Show full event details');
+  });
+
+  it('should render the take action button', () => {
+    (useInvestigateInTimeline as jest.Mock).mockReturnValue({
+      investigateInTimelineActionItems: [{ name: 'test', onClick: jest.fn() }],
+    });
+    const { getByTestId } = render(
+      <TestProviders>
+        <DocumentDetailsContext.Provider value={mockContextValue}>
+          <PreviewPanelFooter />
+        </DocumentDetailsContext.Provider>
+      </TestProviders>
+    );
+    expect(getByTestId(FLYOUT_FOOTER_DROPDOWN_BUTTON_TEST_ID)).toBeInTheDocument();
   });
 
   it('should open document details flyout when clicked', () => {
