+++ x-pack/plugins/security_solution/public/flyout/document_details/preview/footer.tsx
@@ -5,13 +5,14 @@
  * 2.0.
  */
 
-import { EuiLink, EuiFlexGroup, EuiFlexItem } from '@elastic/eui';
+import type { FC } from 'react';
 import React, { useCallback, useMemo } from 'react';
+import { EuiLink, EuiFlyoutFooter, EuiPanel, EuiFlexGroup, EuiFlexItem } from '@elastic/eui';
 import { i18n } from '@kbn/i18n';
 import { useExpandableFlyoutApi } from '@kbn/expandable-flyout';
+import { TakeActionButton } from '../shared/components/take_action_button';
 import { getField } from '../shared/utils';
 import { EventKind } from '../shared/constants/event_kinds';
-import { FlyoutFooter } from '../../shared/components/flyout_footer';
 import { DocumentDetailsRightPanelKey } from '../shared/constants/panel_keys';
 import { useDocumentDetailsContext } from '../shared/context';
 import { PREVIEW_FOOTER_TEST_ID, PREVIEW_FOOTER_LINK_TEST_ID } from './test_ids';
@@ -21,8 +22,8 @@ import { DocumentEventTypes } from '../../../common/lib/telemetry';
 /**
  * Footer at the bottom of preview panel with a link to open document details flyout
  */
-export const PreviewPanelFooter = () => {
-  const { eventId, indexName, scopeId, getFieldsData } = useDocumentDetailsContext();
+export const PreviewPanelFooter: FC = () => {
+  const { eventId, indexName, scopeId, getFieldsData, isPreview } = useDocumentDetailsContext();
   const { openFlyout } = useExpandableFlyoutApi();
   const { telemetry } = useKibana().services;
 
@@ -48,24 +49,36 @@ export const PreviewPanelFooter = () => {
     });
   }, [openFlyout, eventId, indexName, scopeId, telemetry]);
 
+  const fullDetailsLink = useMemo(
+    () => (
+      <EuiLink
+        onClick={openDocumentFlyout}
+        target="_blank"
+        data-test-subj={PREVIEW_FOOTER_LINK_TEST_ID}
+      >
+        <>
+          {i18n.translate('xpack.securitySolution.flyout.preview.openFlyoutLabel', {
+            values: { isAlert },
+            defaultMessage: 'Show full {isAlert, select, true{alert} other{event}} details',
+          })}
+        </>
+      </EuiLink>
+    ),
+    [isAlert, openDocumentFlyout]
+  );
+
+  if (isPreview) return null;
+
   return (
-    <FlyoutFooter data-test-subj={PREVIEW_FOOTER_TEST_ID}>
-      <EuiFlexGroup justifyContent="center">
-        <EuiFlexItem grow={false}>
-          <EuiLink
-            onClick={openDocumentFlyout}
-            target="_blank"
-            data-test-subj={PREVIEW_FOOTER_LINK_TEST_ID}
-          >
-            <>
-              {i18n.translate('xpack.securitySolution.flyout.preview.openFlyoutLabel', {
-                values: { isAlert },
-                defaultMessage: 'Show full {isAlert, select, true{alert} other{event}} details',
-              })}
-            </>
-          </EuiLink>
-        </EuiFlexItem>
-      </EuiFlexGroup>
-    </FlyoutFooter>
+    <EuiFlyoutFooter data-test-subj={PREVIEW_FOOTER_TEST_ID}>
+      <EuiPanel color="transparent">
+        <EuiFlexGroup justifyContent="flexEnd" alignItems="center">
+          <EuiFlexItem grow={false}>{fullDetailsLink}</EuiFlexItem>
+          <EuiFlexItem grow={false}>
+            <TakeActionButton />
+          </EuiFlexItem>
+        </EuiFlexGroup>
+      </EuiPanel>
+    </EuiFlyoutFooter>
   );
 };
