+++ x-pack/plugins/security_solution/public/flyout/document_details/right/test_ids.ts
@@ -9,8 +9,6 @@ import { PREFIX } from '../../shared/test_ids';
 
 export const FLYOUT_BODY_TEST_ID = `${PREFIX}Body` as const;
 export const FLYOUT_FOOTER_TEST_ID = `${PREFIX}Footer` as const;
-export const FLYOUT_FOOTER_DEOPDOEN_BUTTON_TEST_ID =
-  `${FLYOUT_FOOTER_TEST_ID}DropdownButton` as const;
 export const OVERVIEW_TAB_TEST_ID = `${PREFIX}OverviewTab` as const;
 export const TABLE_TAB_TEST_ID = `${PREFIX}TableTab` as const;
 export const JSON_TAB_TEST_ID = `${PREFIX}JsonTab` as const;
