+++ x-pack/plugins/security_solution/public/flyout/document_details/shared/components/take_action_dropdown.test.tsx
@@ -27,7 +27,7 @@ import {
   ALERT_ASSIGNEES_CONTEXT_MENU_ITEM_TITLE,
   ALERT_TAGS_CONTEXT_MENU_ITEM_TITLE,
 } from '../../../../common/components/toolbar/bulk_actions/translations';
-import { FLYOUT_FOOTER_DEOPDOEN_BUTTON_TEST_ID } from '../test_ids';
+import { FLYOUT_FOOTER_DROPDOWN_BUTTON_TEST_ID } from './test_ids';
 
 jest.mock('../../../../common/components/endpoint/host_isolation');
 jest.mock('../../../../common/components/endpoint/responder');
@@ -128,7 +128,7 @@ describe('take action dropdown', () => {
       </TestProviders>
     );
     expect(
-      wrapper.find(`[data-test-subj="${FLYOUT_FOOTER_DEOPDOEN_BUTTON_TEST_ID}"]`).exists()
+      wrapper.find(`[data-test-subj="${FLYOUT_FOOTER_DROPDOWN_BUTTON_TEST_ID}"]`).exists()
     ).toBeTruthy();
   });
 
@@ -139,7 +139,7 @@ describe('take action dropdown', () => {
       </TestProviders>
     );
     expect(
-      wrapper.find(`[data-test-subj="${FLYOUT_FOOTER_DEOPDOEN_BUTTON_TEST_ID}"]`).first().text()
+      wrapper.find(`[data-test-subj="${FLYOUT_FOOTER_DROPDOWN_BUTTON_TEST_ID}"]`).first().text()
     ).toEqual('Take action');
   });
 
@@ -153,7 +153,7 @@ describe('take action dropdown', () => {
         </TestProviders>
       );
       wrapper
-        .find(`button[data-test-subj="${FLYOUT_FOOTER_DEOPDOEN_BUTTON_TEST_ID}"]`)
+        .find(`button[data-test-subj="${FLYOUT_FOOTER_DROPDOWN_BUTTON_TEST_ID}"]`)
         .simulate('click');
     });
     test('should render "Add to existing case"', async () => {
@@ -325,7 +325,7 @@ describe('take action dropdown', () => {
         </TestProviders>
       );
       wrapper
-        .find(`button[data-test-subj="${FLYOUT_FOOTER_DEOPDOEN_BUTTON_TEST_ID}"]`)
+        .find(`button[data-test-subj="${FLYOUT_FOOTER_DROPDOWN_BUTTON_TEST_ID}"]`)
         .simulate('click');
 
       return wrapper;
